split qwen36 moe layer loading
